starter code for decomposation
