+++ processor/src/main/java/my/raptive/url/processor/ProcessStatus.java
@@ -1,0 +1,5 @@
+package my.raptive.url.processor;
+
+public enum ProcessStatus {
+    SUCCESS, FAILED, IN_PROGRESS
+}
