+++ processor/src/main/java/my/raptive/url/processor/process/ScheduledTasks.java
@@ -1,0 +1,31 @@
+package my.raptive.url.processor.process;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.stereotype.Component;
+
+import java.text.SimpleDateFormat;
+import java.util.Date;
+
+@Component
+public class ScheduledTasks {
+
+	@Autowired
+	WebContentProcessService webContentProcessService;
+	private static final SimpleDateFormat dateFormat = new SimpleDateFormat(
+			"MM/dd/yyyy HH:mm:ss");
+
+	@Scheduled(fixedRate = 5000)
+	public void performTask() {
+
+		System.out.println("Regular task performed at "
+				+ dateFormat.format(new Date()));
+
+	}
+
+	@Scheduled(fixedRate = 5000)
+	public void process() {
+		webContentProcessService.process("9048c66e-2723-466c-a752-727061616c97");
+	}
+
+}
