+++ processor/src/test/java/my/raptive/url/processor/UrlProcessServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package my.raptive.url.processor;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UrlProcessServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
