+++ rest/src/main/java/my/raptive/url/rest/ProcessStatus.java
@@ -1,0 +1,5 @@
+package my.raptive.url.rest;
+
+public enum ProcessStatus {
+    SUCCESS, FAILED, IN_PROGRESS
+}
