+++ rest/src/test/java/my/raptive/url/rest/RestAppTests.java
@@ -1,0 +1,13 @@
+package my.raptive.url.rest;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class RestAppTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
