--- src/main/java/my/raptive/urlProcessService/ProcessStatus.java
@@ -1,5 +1,0 @@
-package my.raptive.urlProcessService;
-
-public enum ProcessStatus {
-    SUCCESS, FAILED, IN_PROGRESS
-}
