--- src/main/java/my/raptive/urlProcessService/UrlProcessServiceApplication.java
@@ -1,15 +1,0 @@
-package my.raptive.urlProcessService;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.scheduling.annotation.EnableAsync;
-
-@SpringBootApplication
-@EnableAsync
-public class UrlProcessServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(UrlProcessServiceApplication.class, args);
-	}
-
-}
