--- src/main/java/my/raptive/urlProcessService/repository/UrlProcessInfo.java
@@ -1,42 +1,0 @@
-package my.raptive.urlProcessService.repository;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.experimental.Accessors;
-import my.raptive.urlProcessService.ProcessStatus;
-
-@Getter
-@Setter
-@Accessors(chain = true)
-@NoArgsConstructor
-@Entity
-public class UrlProcessInfo {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private long id;
-
-    @Column(nullable = false)
-    private String processId;
-
-    @Column(nullable = false)
-    private ProcessStatus processStatus;
-
-    @Column(nullable = false)
-    private String url;
-
-    @Column
-    private String title;
-
-    @Column(length = 4000)
-    private String description;
-
-    @Column(length = 4000)
-    private String body;
-
-    @Column(length = 1000)
-    private String errorMessage;
-
-}
