--- src/main/java/my/raptive/urlProcessService/rs/UrlProcessWebService.java
@@ -1,32 +1,0 @@
-package my.raptive.urlProcessService.rs;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.MediaType;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/url/process")
-public class UrlProcessWebService {
-
-	UrlProcessWebServiceHandler urlProcessWebServiceHandler;
-	@Autowired
-	UrlProcessWebService(UrlProcessWebServiceHandler urlProcessWebServiceHandler) {
-		this.urlProcessWebServiceHandler = urlProcessWebServiceHandler;
-	}
-
-	@GetMapping("")
-	public List<WebContentDto> getResult(@RequestParam("processId") String processId) {
-		return urlProcessWebServiceHandler.getResult(processId);
-	}
-
-	@PostMapping(
-			value = "",
-			consumes = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE},
-			produces = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE})
-	public String processUrl(@RequestBody List<String> urls) {
-		return urlProcessWebServiceHandler.processUrl(urls);
-	}
-
-}
