--- src/test/java/my/raptive/urlProcessService/UrlProcessServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package my.raptive.urlProcessService;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class UrlProcessServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
